Docs migration initial commit
The docs have been migrated using the migration script and adapted for the Product docs navigation, which differs from the project one.

The initial version is 5.3.

All the pages have been renamed and moved to the root directory for a better management and all the links have been updated accordingly.

Global variables have been defined in `global-variables.adoc` partial file.

Signed-off-by: Nuno do Carmo nuno.carmo@suse.com